+++ CHANGELOG.md
@@ -4,6 +4,15 @@ All notable changes to this package will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [2.6.1] - 2023-05-31
+### Added
+- Added `SignInWithUnityAsync` public API.
+- Added `LinkWithUnityAsync` public API.
+- Added `UnlinkUnityAsync` public API.
+- Added Unity Player Accounts configuration in Editor settings.
+### Fixed
+- Adding Preserve attributes to api models to prevent issues with code stripping for Oculus
+
 ## [2.5.0] - 2023-05-09
 ### Fixed
 - Properly disposing web request when using player names